requirements/requirements-dev.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
loguru
python-dotenv

# hard dependencies
scikit-learn
scipy
numpy
pandas
umap-learn
pingouin
statsmodels
lightgbm
ipython
ipykernel